don't attempt to delete files that don't exist
Former-commit-id: de5f1f38314489571ae316b7fd54c3c6c351b1a6 [formerly 776a301c944bd9ff60d4581c15a87a3bed77dd07] [formerly f2404d06ed9f7bc1941c63e17d83c1c50751f63f [formerly eda9b78f253534bb3c916d561cc06c87c97ca54f]]
Former-commit-id: 03f27d9f62d5439d93e5613ad532370d10f56328 [formerly fd78178aff7f98f7483fdfcb5ba3f3c26811af59]
Former-commit-id: 015b08f6eda168d8e8a5a03194e0baff990c6f2a

+++ package_release.bat
@@ -47,7 +47,7 @@ if %errorlevel% neq 0 echo "There was an error building QuickFIX/n" && exit /b %
 echo * Built QuickFIX/n.
 
 rem Copy files to temp directory
-rmdir /s /q tmp
+IF EXIST tmp rmdir /s /q tmp
 mkdir tmp
 mkdir tmp\%QF_DIR%
 mkdir tmp\%QF_DIR%\bin
@@ -64,7 +64,7 @@ echo * Copied files to tmp directory.
 
 rem Create ZIP
 set ZIP_NAME=%QF_DIR%.zip
-del /q %ZIP_NAME%
+IF EXIST %ZIP_NAME% del /q %ZIP_NAME%
 ruby scripts\create_zip.rb tmp/%QF_DIR% %ZIP_NAME%
 if %errorlevel% neq 0 echo "There was an error creating QuickFIX/n ZIP: %ZIP_NAME%" && exit /b %errorlevel%
 echo * Created zip.
